# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Consultar todas los productos",
                "description": "Retorna todas los productos",
                "operationId": "3aaeb8ca130d00a7df6a35986370a18f",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No se encontraron productos"
                    },
                    "405": {
                        "description": "Método no permitido"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear productos",
                "description": "Crear un nuevo producto ",
                "operationId": "e5a64f77fdea6674edfd95b565b98556",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "precio",
                                    "stock"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del producto",
                                        "type": "string",
                                        "example": "Producto A"
                                    },
                                    "descripcion": {
                                        "description": "Descripción del producto",
                                        "type": "string",
                                        "example": "Descripción del Producto A"
                                    },
                                    "precio": {
                                        "description": "Precio del producto",
                                        "type": "number",
                                        "format": "float",
                                        "example": "19.99"
                                    },
                                    "stock": {
                                        "description": "Cantidad en stock del producto",
                                        "type": "integer",
                                        "example": "100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/productos/{producto}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Consultar un producto por ID",
                "description": "Retorna un producto específico",
                "operationId": "2d9cc407fd950d8ef46774017f707f58",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "405": {
                        "description": "Método no permitido"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto",
                "description": "Actualiza un producto específico",
                "operationId": "98ce624a5977d6c1e36e37b29605300b",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "nombre",
                                    "descripcion",
                                    "precio",
                                    "stock"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "Debe ser 'PUT' o 'PATCH'",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "nombre": {
                                        "description": "Nombre del producto",
                                        "type": "string",
                                        "example": "Producto B (Actualizado)"
                                    },
                                    "descripcion": {
                                        "description": "Descripción del producto",
                                        "type": "string",
                                        "example": "Descripción del Producto B (Actualizado)"
                                    },
                                    "precio": {
                                        "description": "Precio del producto",
                                        "type": "number",
                                        "format": "float",
                                        "example": "25.50"
                                    },
                                    "stock": {
                                        "description": "Cantidad en stock del producto",
                                        "type": "integer",
                                        "example": "50"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "description": "Elimina un producto específico",
                "operationId": "a008ac3b1d9b6aabbcdd6aed72c6d5a1",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado exitosamente (Sin contenido)"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}